Fix uploader

+++ android/DataCollection/ContextActionLibrary/src/main/java/com/hcifuture/contextactionlibrary/sensor/uploader/Uploader.java
@@ -53,7 +53,7 @@ public class Uploader {
 
     private static final Gson gson = new Gson();
 
-    private Context mContext;
+    private final Context mContext;
 
     private final Lock lock = new ReentrantLock();
     private final Condition compressCondition = lock.newCondition();
@@ -63,17 +63,17 @@ public class Uploader {
     private final PriorityQueue<UploadTask> uploadQueue = new PriorityQueue<>();
     private final int QUEUE_ELEMENT_LIMIT = 10000;
 
-    private ScheduledExecutorService scheduledExecutorService;
-    private List<ScheduledFuture<?>> futureList;
+    private final ScheduledExecutorService scheduledExecutorService;
+    private final List<ScheduledFuture<?>> futureList;
 
     private ScheduledFuture<?> uploadFuture;
     private ScheduledFuture<?> compressFuture;
     private ScheduledFuture<?> localFuture;
 
-    private AtomicBoolean isRunning = new AtomicBoolean(false);
+    private final AtomicBoolean isRunning = new AtomicBoolean(false);
 
-    private String fileFolder;
-    private String zipFolder;
+    private final String fileFolder;
+    private final String zipFolder;
 
     enum UploaderStatus {
         OK,
@@ -102,7 +102,7 @@ public class Uploader {
         futureList.add(localFuture);
     }
 
-    private void stop() {
+    public void stop() {
         if (uploadFuture != null) {
             uploadFuture.cancel(true);
         }
@@ -158,7 +158,8 @@ public class Uploader {
     }
 
     private void compress() {
-        while (isRunning.get()) {
+        List<File> needToDelete = new ArrayList<>();
+        while (!Thread.currentThread().isInterrupted() && isRunning.get()) {
             List<UploadTask> pack = new ArrayList<>();
             try {
                 lock.lock();
@@ -178,38 +179,52 @@ public class Uploader {
                 lock.unlock();
             }
             if (pack.size() > 0) {
-                String zipName = System.currentTimeMillis() + ".zip";
-                String metaName = zipName + ".meta";
                 try {
+                    String zipName = System.currentTimeMillis() + ".zip";
+                    String metaName = zipName + ".meta";
+                    needToDelete.clear();
                     File zipFile = new File(zipFolder + zipName);
                     File metaFile = new File(zipFolder + metaName);
                     FileUtils.makeFile(zipFile);
                     FileUtils.makeFile(metaFile);
-                    ZipOutputStream os = new ZipOutputStream(new FileOutputStream(zipFile));
-                    for (int i = 0; i < pack.size(); i++) {
-                        File file = pack.get(i).getFile();
-                        Log.d(TAG, "[PACK] Compressing " + file.getAbsolutePath());
-                        ZipEntry zipEntry = new ZipEntry(file.getName());
-                        FileInputStream is = new FileInputStream(file);
-                        os.putNextEntry(zipEntry);
-                        int len;
-                        byte[] buffer = new byte[1024 * 1024];
-                        while ((len = is.read(buffer)) != -1) {
-                            os.write(buffer, 0, len);
+                    int packedEntries = 0;
+                    try (ZipOutputStream os = new ZipOutputStream(new FileOutputStream(zipFile))) {
+                        for (int i = 0; i < pack.size(); i++) {
+                            File file = pack.get(i).getFile();
+                            Log.e(TAG, "[PACK] Compressing " + file.getAbsolutePath());
+                            ZipEntry zipEntry = new ZipEntry(file.getName());
+                            // may throw FileNotFoundException
+                            try (FileInputStream is = new FileInputStream(file)) {
+                                os.putNextEntry(zipEntry);
+                                int len;
+                                byte[] buffer = new byte[1024 * 1024];
+                                while ((len = is.read(buffer)) != -1) {
+                                    os.write(buffer, 0, len);
+                                }
+                                os.closeEntry();
+                                // delete already packed files
+                                packedEntries++;
+                                needToDelete.add(file);
+                                needToDelete.add(pack.get(i).getMetaFile());
+                            } catch (Exception e) {
+                                e.printStackTrace();
+                            }
                         }
-                        is.close();
-                        os.closeEntry();
                     }
-                    os.finish();
-                    os.close();
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                        List<TaskMetaBean> metas = pack.stream().map((x) -> x.getMeta().get(0)).collect(Collectors.toList());
-                        FileUtils.writeStringToFile(gson.toJson(metas), metaFile);
-                        pushTask(new UploadTask(zipFile, metaFile, metas), false);
-                    }
-                    for (int i = 0; i < pack.size(); i++) {
-                        FileUtils.deleteFile(pack.get(i).getFile(), "PACK");
-                        FileUtils.deleteFile(pack.get(i).getMetaFile(), "PACK");
+                    Log.e(TAG, "compress: packedEntries: " + packedEntries);
+                    if (packedEntries > 0) {
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                            List<TaskMetaBean> metas = pack.stream().map((x) -> x.getMeta().get(0)).collect(Collectors.toList());
+                            FileUtils.writeStringToFile(gson.toJson(metas), metaFile);
+                            pushTask(new UploadTask(zipFile, metaFile, metas), false);
+                        }
+                        for (int i = 0; i < needToDelete.size(); i++) {
+                            FileUtils.deleteFile(needToDelete.get(i), "PACK");
+                        }
+                    } else {
+                        // pack contains no file, delete .zip and .zip.meta
+                        FileUtils.deleteFile(zipFile, "PACK FAIL");
+                        FileUtils.deleteFile(metaFile, "PACK FAIL");
                     }
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -219,7 +234,7 @@ public class Uploader {
     }
 
     private void upload() {
-        while (isRunning.get()) {
+        while (!Thread.currentThread().isInterrupted() && isRunning.get()) {
             UploadTask task;
             try {
                 lock.lock();
@@ -258,9 +273,11 @@ public class Uploader {
 
                     @Override
                     public void onResponse(@NonNull Call call, @NonNull Response response) {
-                        Log.d(TAG, "Successfully uploaded " + task.getFile().getAbsolutePath());
-                        FileUtils.deleteFile(task.getFile(), "UPLOAD");
-                        FileUtils.deleteFile(task.getMetaFile(), "UPLOAD");
+                        if (response.isSuccessful()) {
+                            Log.d(TAG, "Successfully uploaded " + task.getFile().getAbsolutePath());
+                            FileUtils.deleteFile(task.getFile(), "UPLOAD");
+                            FileUtils.deleteFile(task.getMetaFile(), "UPLOAD");
+                        }
                     }
                 });
             }
@@ -276,32 +293,38 @@ public class Uploader {
     }
 
     private void uploadDirectory(File dir, long timestamp, boolean isZip) {
+        Log.e(TAG, "uploadDirectory: " + dir + " isZip: " + isZip);
         if (dir.exists()) {
             File[] files = dir.listFiles();
             if (files != null) {
                 for (File file : files) {
-                    if (file.isDirectory()) {
-                        uploadDirectory(file, timestamp, isZip);
-                    } else {
-                        File metaFile = new File(file.getAbsolutePath() + ".meta");
-                        if (metaFile.exists()) {
-                            if (isZip) {
-                                Type type = new TypeToken<List<TaskMetaBean>>(){}.getType();
-                                String content = FileUtils.getFileContent(metaFile.getAbsolutePath());
-                                List<TaskMetaBean> meta = gson.fromJson(content, type);
-                                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                                    if (meta.stream().noneMatch(v -> v.getTimestamp() >= timestamp)) {
-                                        pushTask(new UploadTask(file, metaFile, meta), false);
+                    try {
+                        Log.e(TAG, "uploadDirectory: checking file: " + file);
+                        if (file.isDirectory()) {
+                            uploadDirectory(file, timestamp, isZip);
+                        } else {
+                            File metaFile = new File(file.getAbsolutePath() + ".meta");
+                            if (metaFile.exists()) {
+                                if (isZip) {
+                                    Type type = new TypeToken<List<TaskMetaBean>>(){}.getType();
+                                    String content = FileUtils.getFileContent(metaFile.getAbsolutePath());
+                                    List<TaskMetaBean> meta = gson.fromJson(content, type);
+                                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                                        if (meta.stream().noneMatch(v -> v.getTimestamp() >= timestamp)) {
+                                            pushTask(new UploadTask(file, metaFile, meta), false);
+                                        }
                                     }
-                                }
-                            } else {
-                                String content = FileUtils.getFileContent(metaFile.getAbsolutePath());
-                                TaskMetaBean meta = gson.fromJson(content, TaskMetaBean.class);
-                                if (meta.getTimestamp() < timestamp) {
-                                    pushTask(new UploadTask(file, metaFile, meta), true);
+                                } else {
+                                    String content = FileUtils.getFileContent(metaFile.getAbsolutePath());
+                                    TaskMetaBean meta = gson.fromJson(content, TaskMetaBean.class);
+                                    if (meta.getTimestamp() < timestamp) {
+                                        pushTask(new UploadTask(file, metaFile, meta), true);
+                                    }
                                 }
                             }
                         }
+                    } catch (Exception e) {
+                        e.printStackTrace();
                     }
                 }
             }
